+++ code/WorkInProgress/zamujasa.dm
@@ -11,8 +11,8 @@
 	attackby(obj/item/W as obj, mob/user as mob)
 		if (issnippingtool(W))
 			logTheThing("station", user, null, "cut the don't-cut-this wire and got ghosted/disconnected as a result.")
-			//boutput(user, "<span style=\"color:red\">You snip the ca</span>")
-			user.visible_message("[user] nearly snips the cable with \the [W], but suddenly freezes in place just before it cuts!", "<span style=\"color:red\">You snip the ca</span>")
+			//boutput(user, "<span class='alert'>You snip the ca</span>")
+			user.visible_message("[user] nearly snips the cable with \the [W], but suddenly freezes in place just before it cuts!", "<span class='alert'>You snip the ca</span>")
 			var/client/C = user.client
 			user.ghostize()
 			del(C)
@@ -26,7 +26,7 @@
 /obj/item/device/speechtotext
 	name = "dumb microphone"
 	desc = "This is really stupid."
-	icon = 'icons/obj/device.dmi'
+	icon = 'icons/obj/items/device.dmi'
 	icon_state = "mic"
 	item_state = "mic"
 
@@ -80,6 +80,7 @@
 	maptext_width = 96
 
 	New(var/change = 0)
+		..()
 		if (abs(change) < 1)
 			del(src)
 			return
@@ -113,6 +114,7 @@
 	var/bumped = 0
 
 	New(mob/M as mob, msg, style = "")
+		..()
 		for (var/obj/maptext_junk/speech/O in M.vis_contents)
 			if (!istype(O))
 				continue
@@ -251,6 +253,7 @@
 	var/working = 0
 
 	New()
+		..()
 		total_score = world.load_intra_round_value("afterlife_donations")
 		tracker = new /obj/maptext_junk()
 		tracker.pixel_y = 40
@@ -272,7 +275,7 @@
 		if (score == -1)
 			return ..()
 
-		boutput(user, "<span style=\"color:blue\">[src] mulches up [W].</span>")
+		boutput(user, "<span class='notice'>[src] mulches up [W].</span>")
 		user.u_equip(W)
 		W.dropped()
 		mulch_item(W, score)
@@ -352,19 +355,19 @@
 
 	MouseDrop_T(atom/movable/O as mob|obj, mob/user as mob)
 		if (!isliving(user))
-			boutput(user, "<span style=\"color:red\">Excuse me you are dead, get your gross dead hands off that!</span>")
+			boutput(user, "<span class='alert'>Excuse me you are dead, get your gross dead hands off that!</span>")
 			return
 		if (get_dist(user,src) > 1)
-			boutput(user, "<span style=\"color:red\">You need to move closer to [src] to do that.</span>")
+			boutput(user, "<span class='alert'>You need to move closer to [src] to do that.</span>")
 			return
 		if (get_dist(O,src) > 1 || get_dist(O,user) > 1)
-			boutput(user, "<span style=\"color:red\">[O] is too far away to load into [src]!</span>")
+			boutput(user, "<span class='alert'>[O] is too far away to load into [src]!</span>")
 			return
 
 		var/score = 0
 		if (get_item_value(O) != -1)
 			var/MT = start_scoring()
-			user.visible_message("<span style=\"color:blue\">[user] begins quickly stuffing things into [src]!</span>")
+			user.visible_message("<span class='notice'>[user] begins quickly stuffing things into [src]!</span>")
 			var/staystill = user.loc
 
 			for(var/obj/item/P in view(1,user))
@@ -375,22 +378,23 @@
 				score += addscore
 				mulch_item(P, addscore)
 				update_score(MT, score)
-				sleep(1)
+				sleep(0.1 SECONDS)
 
-			boutput(user, "<span style=\"color:blue\">You finish stuffing things into [src]!</span>")
+			boutput(user, "<span class='notice'>You finish stuffing things into [src]!</span>")
 			finish_scoring(MT)
 		else ..()
 
 /obj/death_button/clean_gunsim
 	name = "button that will clean the murderbox"
 	desc = "push this to clean the murderbox and probably not get killed. takes a minute."
-	icon = 'icons/obj/aibots.dmi'
+	icon = 'icons/obj/bots/aibots.dmi'
 	icon_state = "cleanbot1"
 
 	var/area/sim/gunsim/gunsim
 	var/active = 0
 
 	New()
+		..()
 		SPAWN_DBG(0.5 SECONDS)
 			gunsim = locate() in world
 
@@ -406,7 +410,17 @@
 
 		SPAWN_DBG(0)
 			for (var/obj/item/I in gunsim)
-				qdel(I)
+				if(istype(I, /obj/item/device/radio/intercom)) //lets not delete the intercoms inside shall we?
+					continue
+				else
+					qdel(I)
+
+			for (var/atom/S in gunsim)
+				if(istype(S, /obj/storage) || istype(S, /obj/artifact) || istype(S, /obj/critter) || istype(S, /obj/machinery/bot) || istype(S, /obj/decal) || istype(S, /mob/living/carbon/human/tdummy))
+					qdel(S)
+
+
+/*
 			for (var/obj/storage/S in gunsim)
 				qdel(S)
 			for (var/obj/artifact/A in gunsim)
@@ -417,13 +431,40 @@
 				qdel(B)
 			for (var/obj/decal/D in gunsim)
 				qdel(D)
-
+*/
 		SPAWN_DBG(60 SECONDS)
 			active = 0
 			alpha = 255
 			icon_state = "cleanbot1"
 
 
+/obj/death_button/create_dummy
+	name = "Button that creates a test dummy"
+	desc = "click this to create a test dummy"
+	icon = 'icons/mob/human.dmi'
+	icon_state = "ghost"
+	var/active = 0
+	alpha = 255
+
+	attack_hand(mob/user as mob)
+		if (active)
+			boutput(user, "did you already kill the dummy? either way wait a bit!")
+			return
+
+		active = 1
+		alpha = 128
+		boutput(user, "Spawning target dummy, stand by") //no need to be rude
+
+		new /mob/living/carbon/human/tdummy(locate(src.x+1, src.y, src.z))
+		//T.x = src.x + 1 // move it to the right
+
+
+		SPAWN_DBG(10 SECONDS)
+			active = 0
+			alpha = 255
+
+	ex_act(severity)
+		return
 
 
 /proc/fancy_pressure_bar(var/pressure, var/max_pressure, var/width = 300)
@@ -432,7 +473,7 @@
 	var/bar_bg_color = "#000000"
 	var/bar_color = "#00cc00"
 	var/bar_width = clamp(pct, 0, 1)
-	if (pct > 1)
+	if (pct > 1.01)
 		bar_width = clamp((pressure / (max_pressure * 10)), 0, 1)
 		bar_bg_color = "#b00000"
 		bar_color = "#ffff00"
@@ -512,6 +553,8 @@
 	layer = TURF_LAYER + 0.1 // it should basically be part of a turf
 	plane = PLANE_FLOOR // hence, they should be on the same plane!
 
+	ex_act(severity)
+		return
 
 
 /area/football
@@ -564,9 +607,12 @@
 	name = "join"
 	icon = 'icons/effects/mapeditor.dmi'
 	icon_state = "landmark"
+	deleted_on_start = TRUE
+	add_to_landmarks = FALSE
+
 	New()
 		football_spawns[src.name] += src.loc
-		qdel(src)
+		..()
 
 	blue
 		name = "blue"
@@ -622,12 +668,13 @@
 
 
 	New()
-		SubscribeToProcess()
+		..()
 		src.process()
 
 	disposing()
 		UnsubscribeProcess()
-	
+		..()
+
 	process()
 		if (src.last_count != runtime_count)
 			src.last_count = runtime_count
@@ -637,6 +684,256 @@
 			src.maptext_x = -100
 			src.maptext_width = 232
 			src.maptext_y = 34
-		
+
 	ex_act()
 		return
+
+
+
+/obj/machinery/maptext_monitor
+	name = "maptext monitor doodad"
+	desc = "This thing reports the value something else has, automatically! Wow!"
+	icon = null
+	anchored = 2
+	density = 0
+
+	var/datum/monitored = null
+	var/monitored_var = null
+	var/monitored_list = null
+	var/monitored_ref = null
+	var/last_value = null
+	var/display_mode = null
+	var/maptext_prefix = "<span class='c pixel sh'>Value:\n<span class='vga'>"
+	var/maptext_suffix = "</span></span>"
+	var/ding_on_change = 0
+	var/ding_sound = "sound/machines/ping.ogg"
+	var/update_delay = null
+
+	New()
+		..()
+		src.maptext_x = -100
+		src.maptext_width = 232
+		src.maptext_height = 64
+		src.process()
+
+	disposing()
+		UnsubscribeProcess()
+		..()
+
+	process()
+		src.update_monitor()
+		if (src.update_delay)
+			UnsubscribeProcess()
+			SPAWN_DBG(0)
+				while (src.update_delay)
+					src.update_monitor()
+					sleep(update_delay)
+
+				SubscribeToProcess()
+
+	proc/update_monitor()
+		if (src.monitored_ref)
+			var/datum/thing = locate(src.monitored_ref)
+			if (thing)
+				src.monitored = thing
+			src.monitored_ref = null
+
+		if (monitored)
+			if (monitored.pooled || monitored.qdeled)
+				// The thing we were watching was deleted/removed! Welp.
+				monitored = null
+				return
+
+			if (!src.monitored_list && !src.monitored_var)
+				return
+			try
+				var/current_value
+				if (src.monitored_list && !src.monitored_var)
+					var/list/monlist = monitored.vars[src.monitored_list]
+					current_value = monlist.len
+				else if (src.monitored_list)
+					current_value = monitored.vars[src.monitored_list][src.monitored_var]
+				else
+					current_value = monitored.vars[monitored_var]
+
+				if (current_value != last_value)
+					src.maptext = "[maptext_prefix][format_value(current_value)][maptext_suffix]"
+					src.last_value = current_value
+					if (src.ding_on_change)
+						playsound(src, src.ding_sound, 33, 0)
+			catch(var/exception/e)
+				src.maptext = "(Err: [e])"
+
+
+	proc/format_value(var/val)
+		switch (src.display_mode)
+			if ("power")
+				return engineering_notation(val)
+			if ("percent")
+				return (val * 100)
+			if ("temperature")
+				return "[val - T0C]&deg;C"
+			if ("round")
+				return round(val)
+
+		return val
+
+
+	ex_act()
+		return
+
+
+
+
+/obj/overlay/zamujasa/football_wave_timer
+	name = "football wave countdown"
+
+	New()
+		..()
+		src.maptext_x = -100
+		src.maptext_height = 64
+		src.maptext_width = 232
+		src.plane = 100
+		src.anchored = 2
+		src.mouse_opacity = 1
+
+	proc/update_timer(var/num)
+		if (num == -1)
+			src.maptext = ""
+		else
+			src.maptext = {"<span class='c pixel sh'>Next spawn wave in\n<span class='vga'>[round(num)]</span> seconds</span>"}
+
+
+
+/obj/overlay/zamujasa/help_text
+	name = "new player tutorial maptext"
+
+	New()
+		..()
+		src.maptext_x = -100
+		src.maptext_height = 64
+		src.maptext_width = 232
+		src.plane = 100
+		src.anchored = 2
+		src.mouse_opacity = 1
+		src.maptext = {"<div class='c pixel sh' style="background: #00000080;"><strong>-- Welcome to Goonstation! --</strong>
+New? <a href="https://mini.xkeeper.net/ss13/tutorial/" style="color: #8888ff; font-weight: bold;" clss="ol">Click here for a tutorial!</a>
+Ask mentors for help with <strong>F3</strong>
+Contact admins with <strong>F1</strong>
+Read the rules, don't grief, and have fun!</div>"}
+
+
+/obj/overlay/zamujasa/round_start_countdown
+	New()
+		..()
+		if (lobby_titlecard)
+			src.x = lobby_titlecard.x + 13
+			src.y = lobby_titlecard.y + 0
+			src.z = lobby_titlecard.z
+			src.layer = lobby_titlecard.layer + 1
+		else
+			// oops
+			src.x = 7
+			src.y = 2
+			src.z = 1
+			src.layer = 1
+
+		src.maptext = ""
+		src.maptext_width = 320
+		src.maptext_x = -(320 / 2) + 16
+		src.maptext_height = 48
+		src.plane = 100
+
+
+	proc/update_status(var/message)
+		if (message)
+			src.maptext = "<span class='c ol vga vt'>Setting up game...\n<span style='color: #aaaaaa;'>[message]</span></span>"
+		else
+			src.maptext = ""
+
+	timer
+		New()
+			..()
+			if (lobby_titlecard)
+				src.x = lobby_titlecard.x + 13
+				src.y = lobby_titlecard.y + 1
+				src.z = lobby_titlecard.z
+				src.layer = lobby_titlecard.layer + 1
+			else
+				// oops
+				src.x = 7
+				src.y = 1
+				src.z = 1
+				src.layer = 1
+
+			src.maptext = ""
+			src.maptext_width = 320
+			src.maptext_x = -(320 / 2) + 16
+			src.maptext_height = 96
+			src.plane = 100
+
+		proc/update_time(var/time)
+			if (time >= 0)
+				var/timeLeftColor
+				switch (time)
+					if (90 to INFINITY)
+						timeLeftColor = "#33dd33"
+					if (60 to 90)
+						timeLeftColor = "#ffff00"
+					if (30 to 60)
+						timeLeftColor = "#ffb400"
+					if (0 to 30)
+						timeLeftColor = "#ff6666"
+				src.maptext = "<span class='c ol vga vt'>Round begins in<br><span style='color: [timeLeftColor]; font-size: 36px;'>[time]</span></span>"
+			else
+				src.maptext = "<span class='c ol vga vt'>Round begins<br><span style='color: #aaaaaa; font-size: 36px;'>soon</span></span>"
+
+
+
+
+/obj/overlay/inventory_counter
+	name = "inventory amount counter"
+	invisibility = 101
+	plane = PLANE_HUD
+	layer = HUD_LAYER_3
+	appearance_flags = RESET_COLOR | RESET_ALPHA | RESET_TRANSFORM | PIXEL_SCALE
+	var/static/matrix/infinity_matrix = matrix().Turn(90).Translate(18, 1)
+
+	New()
+		..()
+		maptext_width = 64
+		maptext_x = -34
+		maptext_y = 1
+		mouse_opacity = 0
+		maptext = ""
+
+	proc/update_text(var/text)
+		maptext = {"<span class="vb r pixel sh">[text]</span>"}
+		if(src.transform) src.transform = null
+
+	proc/update_number(var/number)
+		if(number == -1)
+			maptext = {"<span class="vb r pixel sh" style="font-size:1.5em;">8</span>"} // pixel font has more symmetric 8, ok?
+			src.transform = infinity_matrix
+			return
+		maptext = {"<span class="vb r xfont sh"[number == 0 ? " style='color: #ff6666;'" : number == -1 ? " style='-ms-transform: rotate(-90deg);'" : ""]>[number == -1 ? "8" : number >= 100000 ? "[round(number / 1000)]K" : round(number)]</span>"}
+		if(src.transform) src.transform = null
+
+	proc/update_percent(var/current, var/maximum)
+		if (!maximum)
+			// no dividing by zero
+			src.update_number(current)
+			return
+		maptext = {"<span class="vb r xfont sh"[current == 0 ? " style='color: #ff6666;'" : ""]>[round(current / maximum * 100)]%</span>"}
+		if(src.transform) src.transform = null
+
+	proc/hide_count()
+		invisibility = 101
+
+	proc/show_count()
+		invisibility = 0
+
+	pooled()
+		src.maptext = ""
+		src.invisibility = 101
+		..()
