+++ code/modules/robotics/bot/buttbot.dm
@@ -3,7 +3,7 @@
 /obj/machinery/bot/buttbot
 	name = "buttbot"
 	desc = "Well I... uh... huh."
-	icon = 'icons/obj/aibots.dmi'
+	icon = 'icons/obj/bots/aibots.dmi'
 	icon_state = "buttbot"
 	layer = 5.0 // Todo layer
 	density = 0
@@ -32,6 +32,10 @@
 	name = "robuttbot"
 	icon_state = "cyberbuttbot"
 
+/obj/machinery/bot/buttbot/text2speech
+	text2speech = 1
+
+
 /obj/machinery/bot/buttbot/process()
 	if (prob(10) && src.on == 1)
 		SPAWN_DBG(0)
@@ -51,7 +55,7 @@
 		if (user)
 			user.show_text("You short out the vocal emitter on [src].", "red")
 		SPAWN_DBG(0)
-			src.visible_message("<span style=\"color:red\"><B>[src] buzzes oddly!</B></span>")
+			src.visible_message("<span class='alert'><B>[src] buzzes oddly!</B></span>")
 			playsound(src.loc, "sound/misc/extreme_ass.ogg", 50, 1)
 		src.emagged = 1
 		return 1
@@ -69,22 +73,17 @@
 	if (istype(W, /obj/item/card/emag))
 		//Do not hit the buttbot with the emag tia
 	else
-		src.visible_message("<span style=\"color:red\">[user] hits [src] with [W]!</span>")
-		switch(W.damtype)
-			if("fire")
-				src.health -= W.force * 0.5
-			if("brute")
-				src.health -= W.force * 0.5
-			else
+		src.visible_message("<span class='alert'>[user] hits [src] with [W]!</span>")
+		src.health -= W.force * 0.5
 		if (src.health <= 0)
 			src.explode()
 
 /obj/machinery/bot/buttbot/hear_talk(var/mob/living/carbon/speaker, messages, real_name, lang_id)
 	if(!messages || !src.on)
 		return
-	var/m_id = (lang_id == "english" || lang_id == "") ? messages[1] : messages[2]
+	var/message = (lang_id == "english" || lang_id == "") ? messages[1] : messages[2]
 	if(prob(25))
-		var/list/speech_list = splittext(messages[m_id], " ")
+		var/list/speech_list = splittext(message, " ")
 		if(!speech_list || !speech_list.len)
 			return
 
@@ -104,10 +103,10 @@
 	return src.explode()
 
 /obj/machinery/bot/buttbot/explode()
+	if(src.exploding) return
+	src.exploding = 1
 	src.on = 0
-	src.visible_message("<span style=\"color:red\"><B>[src] blows apart!</B></span>")
-	var/datum/effects/system/spark_spread/s = unpool(/datum/effects/system/spark_spread)
-	s.set_up(3, 1, src)
-	s.start()
+	src.visible_message("<span class='alert'><B>[src] blows apart!</B></span>")
+	elecflash(src, radius=1, power=3, exclude_center = 0)
 	qdel(src)
 	return
